eric's work so far

+++ app/forms.py
@@ -1,10 +1,10 @@
 from flask_wtf import FlaskForm
-from wtforms import DecimalField, SubmitField
+from wtforms import FloatField, SubmitField
 from wtforms.validators import DataRequired
 
 class RunForm(FlaskForm):
-    distance = DecimalField('Distance (mi)', validators=[DataRequired()])
-    pace = DecimalField('Minutes per Mile', validators=[DataRequired()])
-    latitude = DecimalField('Latitude', validators=[DataRequired()])
-    longitude = DecimalField('Longitude', validators=[DataRequired()])
+    distance = FloatField('Distance (mi)', validators=[DataRequired()])
+    pace = FloatField('Minutes per Mile', validators=[DataRequired()])
+    latitude = FloatField('Latitude', validators=[DataRequired()])
+    longitude = FloatField('Longitude', validators=[DataRequired()])
     submit = SubmitField('Submit run data')
